backend/requirements.txt
# Quantara backend — runtime dependencies (install: pip install -r requirements.txt)
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic>=2.0
pydantic-settings>=2.0
psycopg[binary]>=3.1
alembic>=1.13
bcrypt>=4.1
pyjwt>=2.8
httpx>=0.27
pypdf>=5.0
